Refactor to viewmodel style
