+++ app/imports/startup/client/bindings.js
@@ -1,0 +1,3 @@
+import { ViewModel } from 'meteor/manuel:viewmodel';
+
+ViewModel.addAttributeBinding(['src','href','readonly','placeholder']);
