+++ app/imports/startup/client/index.js
@@ -1,4 +1,6 @@
+import './mixins.js';
 import './routes.js';
 import './useraccounts.js';
 import './userpresence.js';
-import './helpers.js';
+import './bindings.js';
+
