+++ app/imports/ui/components/dashboard/user-stats/user-stats.js
@@ -1,4 +1,5 @@
 Template.UserStats.viewmodel({
+  mixin: ['user'],
   usersOnline() {
     return Meteor.users.find({status: 'online'});
   }
