+++ app/imports/ui/components/userdirectory/users-list.js
@@ -1,6 +1,7 @@
 import { Template } from 'meteor/templating';
 
 Template.UsersList.viewmodel({
+  mixin: ['user'],
   isActiveUser(userId) {
     return this.parent().activeUser() === userId;
   }
