+++ app/imports/ui/pages/user-directory-page.js
@@ -18,6 +18,9 @@ Template.UserDirectoryPage.viewmodel({
       }
     }
   },
+  user() {
+    return Meteor.users.findOne({ _id: this.activeUser() })
+  },
   organizationUsers() {
     const userIds = getOrganizationUsers();
